Separate unit-tests and test-tube tests for merkle-incentives contract (#681)
