+++ smart-contracts/contracts/merkle-incentives/Cargo.toml
@@ -11,10 +11,9 @@ crate-type = ["cdylib", "rlib"]
 name = "schema"
 
 [features]
-# for more explicit tests, cargo test --features=backtraces
 backtraces = ["cosmwasm-std/backtraces"]
-# use library feature to disable all instantiate/execute/query exports
 library = []
+test-tube = ["osmosis-test-tube"]
 
 [dependencies]
 base64 = { workspace = true }
@@ -26,8 +25,5 @@ schemars = { workspace = true }
 serde = { workspace = true }
 serde-json-wasm = { workspace = true }
 thiserror = { workspace = true }
-
 rs_merkle = { version = "1.4.2", default-features = false }
-
-[dev-dependencies]
-osmosis-test-tube = { workspace = true }
+osmosis-test-tube = { workspace = true, optional = true }
