+++ smart-contracts/contracts/merkle-incentives/src/lib.rs
@@ -6,7 +6,4 @@ pub mod incentives;
 pub mod msg;
 pub mod state;
 
-#[cfg(test)]
-mod test_tube;
-
 pub use crate::error::ContractError;
