--- smart-contracts/contracts/merkle-incentives/src/test_tube/mod.rs
@@ -1,2 +1,0 @@
-mod initialize;
-mod merkle;
